+++ init.lua
@@ -1,438 +1,80 @@
--- ==== WAYWALL ====
-local waywall = require("waywall")
-local helpers = require("waywall.helpers")
+-- ==== WAYWALL GENERIC CONFIG ====
+local main = require("main")
 
--- ==== RESOURCES ====
-local waywall_config_path = os.getenv("HOME") .. "/.config/waywall/"
-local bg_path = waywall_config_path .. "resources/background.png"
-local tall_overlay_path = waywall_config_path .. "resources/overlay_tall.png"
-local thin_overlay_path = waywall_config_path .. "resources/overlay_thin.png"
-local wide_overlay_path = waywall_config_path .. "resources/overlay_wide.png"
+local remaps = require("remaps")
 
-local pacem_path = waywall_config_path .. "resources/paceman-tracker-0.7.2.jar"
-local nb_path = waywall_config_path .. "resources/Ninjabrain-Bot-1.5.2.jar"
-local overlay_path = waywall_config_path .. "resources/measuring_overlay.png"
-local stretched_overlay_path = waywall_config_path .. "resources/stretched_overlay.png"
+local cfg = {
+    debug_text = true,
 
--- ==== INITS ====
-local remaps_active = true
-local rebind_text = nil
-local thin_active = false
-local keybinds_text = nil
-local debug_text1 = nil
-local debug_text2 = nil
-local debug_text3 = nil
-local debug_text4 = nil
-local debug_text = "Press Shift + I to show keybinds.\n\n" ..
-    "Look at the Github's README for a guide to config.\n\n" ..
-    "disable this message by setting\n" ..
-    "\'debug_text\' to false in ~/.config/waywall/config.lua\n"
+    -- ==== LOOKS ====
+    resolution = { 1920, 1080 },
 
--- ==== CONFIG TABLE ====
+    bg_col = "#000000",
+    toggle_bg_picture = false,
+    text_col = "#FFFFFF",
+    pie_chart_1 = "#EC6E4E",
+    pie_chart_2 = "#46CE66",
+    pie_chart_3 = "#E446C4",
 
-return function(cfg, remaps)
-    local keyboard_remaps = remaps.remapped_kb
-    local other_remaps = remaps.normal_kb
+    ninbot_anchor = {
+        position = "topright", -- topleft, top, topright, left, right, bottomleft, bottomright
+        x = 0,
+        y = 130,               -- offset
+    },
+    ninbot_opacity = 1,        -- 0 to 1
 
-    local config = {
-        input = {
-            layout = (cfg.xkb_config.enabled and cfg.xkb_config.layout) or nil,
-            rules = (cfg.xkb_config.enabled and cfg.xkb_config.rules) or nil,
-            variant = (cfg.xkb_config.enabled and cfg.xkb_config.variant) or nil,
-            options = (cfg.xkb_config.enabled and cfg.xkb_config.options) or nil,
 
-            repeat_rate = 40,
-            repeat_delay = 300,
-            remaps = keyboard_remaps,
-            sensitivity = (cfg.sens_change.enabled and cfg.sens_change.normal) or 1.0,
-            confine_pointer = false,
-        },
-        theme = {
-            background = cfg.bg_col,
-            background_png = cfg.toggle_bg_picture and bg_path or nil,
-            ninb_anchor = cfg.ninbot_anchor,
-            ninb_opacity = cfg.ninbot_opacity,
-        },
-        experimental = {
-            debug = false,
-            jit = false,
-            tearing = false,
-        },
-        window = {
-            fullscreen_width = cfg.resolution[1],
-            fullscreen_height = cfg.resolution[2],
-        }
-    }
+    -- ==== ALTERNATIVE RESOLUTIONS ====
+    thin_res = { 340, 1080 },
+    wide_res = { 1920, 300 },
+    tall_res = { 384, 16384 },
 
-    -- ==== TOOLS ====
-    local is_ninb_running = function()
-        local handle = io.popen("ps aux | grep '[N]injabrain-Bot.*\\.jar'")
-        local result = handle:read("*l")
-        handle:close()
-        return result ~= nil
-    end
-    local is_pacem_running = function()
-        local handle = io.popen("ps aux | grep '[p]aceman-tracker.*\\.jar'")
-        local result = handle:read("*l")
-        handle:close()
-        return result ~= nil
-    end
 
     -- ==== MIRRORS ====
-    -- colors
-    local pie_colors = {
-        { input = "#EC6E4E", output = cfg.pie_chart_1 },
-        { input = "#46CE66", output = cfg.pie_chart_2 },
-        { input = "#CC6C46", output = cfg.pie_chart_2 },
-        { input = "#464C46", output = cfg.pie_chart_2 },
-        { input = "#E446C4", output = cfg.pie_chart_3 }
-    }
-    local percentage_colors = {
-        { input = "#E96D4D", output = cfg.percentages_match_text and cfg.text_col or cfg.pie_chart_1 },
-        { input = "#45CB65", output = cfg.percentages_match_text and cfg.text_col or cfg.pie_chart_2 },
-    }
+    e_count = { enabled = true, x = 1340, y = 300, size = 5, colorkey = false },
 
-    -- thin mirrors
-    if cfg.e_count.enabled then
-        helpers.res_mirror(
-            {
-                src = { x = 13, y = 37, w = 37, h = 9 },
-                dst = { x = cfg.e_count.x, y = cfg.e_count.y, w = 37 * cfg.e_count.size, h = 9 * cfg.e_count.size },
-                depth = 2,
-                color_key = cfg.e_count.colorkey and {
-                    input = "#DDDDDD",
-                    output = cfg.text_col,
-                } or nil,
-            },
-            cfg.thin_res[1], cfg.thin_res[2]
-        )
-        helpers.res_mirror(
-            {
-                src = { x = 13, y = 37, w = 37, h = 9 },
-                dst = { x = cfg.e_count.x, y = cfg.e_count.y, w = 37 * cfg.e_count.size, h = 9 * cfg.e_count.size },
-                depth = 2,
-                color_key = cfg.e_count.colorkey and {
-                    input = "#DDDDDD",
-                    output = cfg.text_col,
-                } or nil,
-            },
-            cfg.tall_res[1], cfg.tall_res[2]
-        )
-    end
+    thin_pie = { enabled = true, x = 1200, y = 400, size = 4, colorkey = false }, -- Turning off colorkeying also maintains the original pie chart's dimensions and shows the percentages
+    tall_pie = { enabled = true, x = 1200, y = 400, size = 4, colorkey = false }, -- Leave same as thin for seamlessness
 
-    if cfg.thin_pie.enabled then
-        if cfg.thin_pie.colorkey then
-            for _, ck in ipairs(pie_colors) do
-                helpers.res_mirror(
-                    {
-                        src = { x = cfg.thin_res[1] - 340, y = cfg.thin_res[2] - 406, w = 340, h = 178 },
-                        dst = { x = cfg.thin_pie.x, y = cfg.thin_pie.y, w = 420 * cfg.thin_pie.size / 4, h = 423 * cfg.thin_pie.size / 4 },
-                        depth = 2,
-                        color_key = ck,
-                    },
-                    cfg.thin_res[1], cfg.thin_res[2]
-                )
-            end
-        else
-            helpers.res_mirror(
-                {
-                    src = { x = cfg.thin_res[1] - 340, y = cfg.thin_res[2] - 406, w = 340, h = 221 },
-                    dst = { x = cfg.thin_pie.x, y = cfg.thin_pie.y, w = 420 * cfg.thin_pie.size / 4, h = 273 * cfg.thin_pie.size / 4 },
-                    depth = 2,
-                },
-                cfg.thin_res[1], cfg.thin_res[2]
-            )
-        end
-    end
+    thin_percent = { enabled = false, x = 1300, y = 850, size = 6 },
+    tall_percent = { enabled = false, x = 1300, y = 850, size = 6 }, -- Leave same as thin for seamlessness
+    percentages_match_text = false,                                  -- Enabling this makes the percentages match the text color rather than the pie colors
 
-    if cfg.thin_percent.enabled then
-        for _, ck in ipairs(percentage_colors) do
-            helpers.res_mirror(
-                {
-                    src = { x = cfg.thin_res[1] - 93, y = cfg.thin_res[2] - 221, w = 33, h = 25 },
-                    dst = { x = cfg.thin_percent.x, y = cfg.thin_percent.y, w = 33 * cfg.thin_percent.size, h = 25 * cfg.thin_percent.size },
-                    depth = 3,
-                    color_key = ck,
-                },
-                cfg.thin_res[1], cfg.thin_res[2]
-            )
-        end
-    end
 
-    -- tall mirrors
-    if cfg.tall_pie.enabled then
-        if cfg.tall_pie.colorkey then
-            for _, ck in ipairs(pie_colors) do
-                helpers.res_mirror(
-                    {
-                        src = { x = 44, y = 15978, w = 340, h = 178 },
-                        dst = { x = cfg.tall_pie.x, y = cfg.tall_pie.y, w = 420 * cfg.tall_pie.size / 4, h = 423 * cfg.tall_pie.size / 4 },
-                        depth = 2,
-                        color_key = ck,
-                    },
-                    cfg.tall_res[1], cfg.tall_res[2]
-                )
-            end
-        else
-            helpers.res_mirror(
-                {
-                    src = { x = 44, y = 15978, w = 340, h = 221 },
-                    dst = { x = cfg.tall_pie.x, y = cfg.tall_pie.y, w = 420 * cfg.tall_pie.size / 4, h = 273 * cfg.tall_pie.size / 4 },
-                    depth = 2,
-                },
-                cfg.tall_res[1], cfg.tall_res[2]
-            )
-        end
-    end
+    measuring_window = { enabled = true, x = 30, y = 340, size = 10 },
+    stretched_measure = false,
 
-    if cfg.tall_percent.enabled then
-        for _, ck in ipairs(percentage_colors) do
-            helpers.res_mirror(
-                {
-                    src = { x = 291, y = 16163, w = 33, h = 25 },
-                    dst = { x = cfg.tall_percent.x, y = cfg.tall_percent.y, w = 33 * cfg.tall_percent.size, h = 25 * cfg.tall_percent.size },
-                    depth = 3,
-                    color_key = ck,
-                },
-                cfg.tall_res[1], cfg.tall_res[2]
-            )
-        end
-    end
-
-    helpers.res_mirror(
-        {
-            src = cfg.stretched_measure
-                and { x = (cfg.tall_res[1] - 30) / 2, y = (cfg.tall_res[2] - 580) / 2, w = 30, h = 580 }
-                or { x = (cfg.tall_res[1] - 60) / 2, y = (cfg.tall_res[2] - 580) / 2, w = 60, h = 580 },
-            dst = { x = cfg.measuring_window.x, y = cfg.measuring_window.y, w = 70 * cfg.measuring_window.size, h = 40 * cfg.measuring_window.size },
-            depth = 2,
-        },
-        cfg.tall_res[1], cfg.tall_res[2]
-    )
-
-    -- ==== IMAGES ====
-    helpers.res_image(
-        cfg.stretched_measure and stretched_overlay_path or overlay_path,
-        {
-            dst = { x = cfg.measuring_window.x, y = cfg.measuring_window.y, w = 70 * cfg.measuring_window.size, h = 40 * cfg.measuring_window.size },
-            depth = 3,
-        },
-        cfg.tall_res[1], cfg.tall_res[2]
-    )
-    helpers.res_image(
-        tall_overlay_path,
-        {
-            dst = { x = 0, y = 0, w = cfg.resolution[1], h = cfg.resolution[2] },
-            depth = 1,
-        },
-        cfg.tall_res[1], cfg.tall_res[2]
-    )
-    helpers.res_image(
-        wide_overlay_path,
-        {
-            dst = { x = 0, y = 0, w = cfg.resolution[1], h = cfg.resolution[2] },
-            depth = 1,
-        },
-        cfg.wide_res[1], cfg.wide_res[2]
-    )
-    helpers.res_image(
-        thin_overlay_path,
-        {
-            dst = { x = 0, y = 0, w = cfg.resolution[1], h = cfg.resolution[2] },
-            depth = 1,
-        },
-        cfg.thin_res[1], cfg.thin_res[2]
-    )
-
-    -- ==== DEBUG TEXT ====
-    waywall.listen("load", function()
-        if cfg.debug_text then
-            debug_text1 = waywall.text(debug_text,
-                { x = 10, y = 10, color = "#FFFF00", size = 3 })
-            debug_text2 = waywall.text(debug_text,
-                { x = 11, y = 11, color = "#FFFF00", size = 3 })
-            debug_text3 = waywall.text(debug_text,
-                { x = 13, y = 13, color = "#000000", size = 3 })
-            debug_text4 = waywall.text(debug_text,
-                { x = 14, y = 14, color = "#000000", size = 3 })
-        end
-    end)
-
-    -- ==== RESIZING STATES ====
-    local thin_enable = function()
-        thin_active = true
-        if cfg.sens_change.enabled then
-            waywall.set_sensitivity(cfg.sens_change.normal)
-        end
-    end
-    local tall_enable = function()
-        if cfg.sens_change.enabled and not thin_active then
-            waywall.set_sensitivity(cfg.sens_change.tall)
-        end
-        thin_active = false
-    end
-    local wide_enable = function()
-        if cfg.sens_change.enabled then
-            waywall.set_sensitivity(cfg.sens_change.normal)
-        end
-        thin_active = false
-    end
-    local res_disable = function()
-        if cfg.sens_change.enabled then
-            waywall.set_sensitivity(cfg.sens_change.normal)
-        end
-        thin_active = false
-    end
-
-    -- ==== RESOLUTIONS ====
-    local make_res = function(width, height, enable, disable)
-        return function()
-            local active_width, active_height = waywall.active_res()
-
-            if active_width == width and active_height == height then
-                if cfg.enable_resize_animations then
-                    os.execute('echo "0x0" > ~/.resetti_state')
-                    waywall.sleep(17)
-                end
-                waywall.set_resolution(0, 0)
-                disable()
-            else
-                if cfg.enable_resize_animations then
-                    os.execute(string.format('echo "%dx%d" > ~/.resetti_state', width, height))
-                    waywall.sleep(17)
-                end
-                waywall.set_resolution(width, height)
-                enable()
-            end
-        end
-    end
-
-    local resolutions = {
-        thin = make_res(cfg.thin_res[1], cfg.thin_res[2], thin_enable, res_disable),
-        tall = make_res(cfg.tall_res[1], cfg.tall_res[2], tall_enable, res_disable),
-        wide = make_res(cfg.wide_res[1], cfg.wide_res[2], wide_enable, res_disable),
-    }
-
-    local function resize_helper(mode, run, ingame_only)
-        local resize = function()
-            if not remaps_active then
-                return false
-            end
-            if mode.f3_safe and waywall.get_key("F3") then
-                return false
-            end
-            return run()
-        end
-
-        if ingame_only then
-            return helpers.ingame_only(resize)
-        end
-
-        return resize
-    end
 
     -- ==== KEYBINDS ====
-    config.actions = {
+    -- resolution changes
+    thin = { key = "*-Alt_L", f3_safe = false, ingame_only = true },
+    wide = { key = "*-B", f3_safe = true, ingame_only = true },
+    tall = { key = "*-F4", f3_safe = false, ingame_only = false },
 
-        [cfg.thin.key] = resize_helper(cfg.thin, resolutions.thin, cfg.thin.ingame_only),
-        [cfg.wide.key] = resize_helper(cfg.wide, resolutions.wide, cfg.wide.ingame_only),
-        [cfg.tall.key] = resize_helper(cfg.tall, resolutions.tall, cfg.tall.ingame_only),
+    -- startup actions
+    toggle_fullscreen_key = "Shift-O",
+    launch_paceman_key = "Shift-P",
 
-        [cfg.toggle_fullscreen_key] = waywall.toggle_fullscreen,
-
-        [cfg.launch_paceman_key] = function()
-            if is_pacem_running() then
-                print("Paceman Already Running")
-            else
-                waywall.exec("java -jar " .. pacem_path .. " --nogui")
-                print("Paceman Running")
-            end
-        end,
-
-        [cfg.toggle_ninbot_key] = function()
-            if not is_ninb_running() then
-                waywall.exec("java -Dawt.useSystemAAFontSettings=on -jar " .. nb_path)
-                waywall.show_floating(true)
-            else
-                helpers.toggle_floating()
-            end
-        end,
-
-        [cfg.toggle_remaps_key] = function()
-            if rebind_text then
-                rebind_text:close()
-                rebind_text = nil
-            end
-            if remaps_active then
-                remaps_active = false
-                waywall.set_remaps(other_remaps)
-
-                if cfg.xkb_config.enabled then
-                    waywall.set_keymap({
-                        layout = nil,
-                        rules = nil,
-                        variant = nil,
-                        options = nil,
-                    })
-                end
-
-                rebind_text = waywall.text(cfg.remaps_text_config.text,
-                    {
-                        x = cfg.remaps_text_config.x,
-                        y = cfg.remaps_text_config.y,
-                        color = cfg.remaps_text_config.color,
-                        size = cfg.remaps_text_config.size
-                    })
-            else
-                remaps_active = true
-                waywall.set_remaps(keyboard_remaps)
-
-                if cfg.xkb_config.enabled then
-                    waywall.set_keymap({
-                        layout = cfg.xkb_config.layout,
-                        rules = cfg.xkb_config.rules,
-                        variant = cfg.xkb_config.variant,
-                        options = cfg.xkb_config.options
-                    })
-                end
-            end
-        end,
-
-        ["Shift-I"] = function()
-            if keybinds_text then
-                keybinds_text:close()
-                keybinds_text = nil
-                return false
-            end
-            keybinds_text = waywall.text(
-                "KEYBINDS:\n" ..
-                "Thin = " .. cfg.thin.key .. "\n" ..
-                "Wide = " .. cfg.wide.key .. "\n" ..
-                "Tall = " .. cfg.tall.key .. "\n" ..
-                "Toggle Ninbot = " .. cfg.toggle_ninbot_key .. "\n" ..
-                "Launch paceman = " .. cfg.launch_paceman_key .. "\n" ..
-                "Fullscreen = " .. cfg.toggle_fullscreen_key .. "\n" ..
-                "Chat Mode = " .. cfg.toggle_remaps_key .. "\n" ..
-                "  "
+    -- during game actions
+    toggle_ninbot_key = "*-apostrophe",
+    toggle_remaps_key = "Insert",
 
 
-                ,
-                { x = 10, y = 10, color = "#FFFFFF", size = 3 })
-            if debug_text1 and debug_text2 and debug_text3 and debug_text4 then
-                debug_text1:close()
-                debug_text1 = nil
-                debug_text2:close()
-                debug_text2 = nil
-                debug_text3:close()
-                debug_text3 = nil
-                debug_text4:close()
-                debug_text4 = nil
-            end
-        end
+    -- ==== KEYBOARD ====
+    xkb_config = {     -- set any setting to nil if unwanted
+        enabled = false,
+        layout = "mc", -- ~/.config/xkb/symbols/mc
+        rules = nil,   -- ~/.config/xkb/rules/...
+        variant = "basic",
+        options = "caps:none",
+    },
+    remaps_text_config = { text = "chat mode", x = 100, y = 100, size = 2, color = "#000000" },
 
-    }
 
-    require("extras")(config)
+    -- ==== MISC ====
+    sens_change = { enabled = false, normal = 1.0, tall = 0.1 }, -- make sure raw input is off
+    enable_resize_animations = false,
 
-    return config
-end
+}
+
+return main(cfg, remaps)
